# perf: allow single typescript project instance