允许单个 Project 实例，在我的项目里性能有 2 倍提升。

```
npm run build-e2e  43.92s user 1.68s system 162% cpu 27.983 total
npm run build-e2e  19.99s user 0.72s system 157% cpu 13.183 total
```

两个遗留 todo：

1. 先暴力地把所有 options 都灌进去 compile 方法，constructor 只接受 compilerOptions。后续可以把更多静态的参数挪到 constructor 里面。
2. features 下的 .ts 都没有 export 因此 TypeScript 认为它们都是 global 模块，放在同一个 project 里会冲突。后续做成 module 后再优化其性能。