@cxtom 2019-12-12
:tada: This PR is included in version 0.19.8 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/ts2php)
- [GitHub release](https://github.com/searchfe/ts2php/releases/tag/v0.19.8)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: